Set up Kori as Razor class library with full reference to ASP.NET core framework (for setup extensions and culture access)
