+++ Kori/ServiceCollectionExtensions.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using System.Globalization;
 
 namespace Kori;
 
@@ -6,15 +8,28 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddKori(this IServiceCollection services)
     {
-        var apiBaseUrl = configuration?["IbisApi"] ?? "https://ibis.chat/";
+        //var apiBaseUrl = configuration?["IbisApi"] ?? "https://ibis.chat/";
 
-        var client = services.AddHttpClient<KoriClient>(client =>
-        {
-            client.BaseAddress = new Uri(apiBaseUrl);
-            client.DefaultRequestVersion = new Version(2, 0);
-        });
+        //var client = services.AddHttpClient<KoriClient>(client =>
+        //{
+        //    client.BaseAddress = new Uri(apiBaseUrl);
+        //    client.DefaultRequestVersion = new Version(2, 0);
+        //});
 
-        services.AddScoped<IbisTranslator>();
+        //services.AddScoped<IbisTranslator>();
         return services;
+    }
+
+    public static IApplicationBuilder UseKori(this IApplicationBuilder app)
+    {
+        var supportedCultures = CultureInfo.GetCultures(CultureTypes.AllCultures)
+            .Select(x => x.Name)
+            .ToArray();
+
+        app.UseRequestLocalization(options => options
+            .AddSupportedCultures(supportedCultures)
+            .AddSupportedUICultures(supportedCultures));
+
+        return app;
     }
 }
