+++ Sparc.sln
@@ -5,15 +5,13 @@ VisualStudioVersion = 17.2.32516.85
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sparc.Coop", "SparcFeatures\Sparc.Coop.csproj", "{40431EDB-80E4-4D29-8A82-E96C3C626C4E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kori", "Kori\Kori.csproj", "{70854DC3-8874-471C-8FB7-C54DB533760A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kori", "Kori\Kori.csproj", "{70854DC3-8874-471C-8FB7-C54DB533760A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ibis", "..\ibis8\Ibis.Features\Ibis.csproj", "{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sparc.Ibis", "..\ibis8\Sparc.Ibis\Sparc.Ibis.csproj", "{228E780C-89F9-4D06-9D62-D95430D8DE10}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_Ibis", "_Ibis", "{6E7CBE83-E569-4801-A4CA-E8D520912B92}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_Site", "_Site", "{04A778CE-534B-43CE-B2AA-239BBA74E029}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,7 +38,6 @@ Global
 		{70854DC3-8874-471C-8FB7-C54DB533760A}.Release|x64.ActiveCfg = Release|Any CPU
 		{70854DC3-8874-471C-8FB7-C54DB533760A}.Release|x64.Build.0 = Release|Any CPU
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Debug|x64.ActiveCfg = Debug|x64
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Debug|x64.Build.0 = Debug|x64
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -48,7 +45,6 @@ Global
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Release|x64.ActiveCfg = Release|x64
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612}.Release|x64.Build.0 = Release|x64
 		{228E780C-89F9-4D06-9D62-D95430D8DE10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{228E780C-89F9-4D06-9D62-D95430D8DE10}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{228E780C-89F9-4D06-9D62-D95430D8DE10}.Debug|x64.ActiveCfg = Debug|Any CPU
 		{228E780C-89F9-4D06-9D62-D95430D8DE10}.Debug|x64.Build.0 = Debug|Any CPU
 		{228E780C-89F9-4D06-9D62-D95430D8DE10}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -60,7 +56,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{40431EDB-80E4-4D29-8A82-E96C3C626C4E} = {04A778CE-534B-43CE-B2AA-239BBA74E029}
 		{98C8DD35-CF11-470B-AFD7-B6AD32BC6612} = {6E7CBE83-E569-4801-A4CA-E8D520912B92}
 		{228E780C-89F9-4D06-9D62-D95430D8DE10} = {6E7CBE83-E569-4801-A4CA-E8D520912B92}
 	EndGlobalSection
