+++ SparcFeatures/Community/CommunityCta.razor
@@ -41,7 +41,7 @@
     {
         if (EmailAddress != null)
         {
-            await RegisterForCommunity.ExecuteAsync(new(EmailAddress));
+            //await RegisterForCommunity.ExecuteAsync(new(EmailAddress));
         }
     }
 }
