+++ SparcFeatures/Community/RegisterForCommunity.cs
@@ -1,21 +1,21 @@
-﻿using Sparc.Kernel;
-using Sparc.Notifications.Twilio;
+﻿//using Sparc.Kernel;
+//using Sparc.Notifications.Twilio;
 
-namespace Sparc.Coop.Community;
+//namespace Sparc.Coop.Community;
 
-public record RegisterForCommunityRequest(string Email);
-public class RegisterForCommunity : Feature<RegisterForCommunityRequest, bool>
-{
-    TwilioService Twilio { get; set; }
+//public record RegisterForCommunityRequest(string Email);
+//public class RegisterForCommunity : Feature<RegisterForCommunityRequest, bool>
+//{
+//    TwilioService Twilio { get; set; }
 
-    public RegisterForCommunity(TwilioService twilio)
-    {
-        Twilio = twilio;
-    }
+//    public RegisterForCommunity(TwilioService twilio)
+//    {
+//        Twilio = twilio;
+//    }
 
-    public override async Task<bool> ExecuteAsync(RegisterForCommunityRequest request)
-    {
-        await Twilio.AddContactAsync(request.Email, "689a1a54-cef8-4327-a04d-abd6a8787dd9");
-        return true;
-    }
-}
+//    public override async Task<bool> ExecuteAsync(RegisterForCommunityRequest request)
+//    {
+//        await Twilio.AddContactAsync(request.Email, "689a1a54-cef8-4327-a04d-abd6a8787dd9");
+//        return true;
+//    }
+//}
