+++ SparcFeatures/Html.scss
@@ -1,0 +1,22 @@
+body {
+    position: relative;
+    margin: 0;
+    overflow-x: hidden;
+    background: var(--background);
+    font-family: var(--encode_semiexpanded);
+    font-weight: 400;
+    color: var(--white);
+}
+
+p {
+    font-family: var(--encode_semiexpanded);
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 160%;
+}
+
+a, button, label, input, select, div, span, li {
+    &:hover {
+        cursor: none;
+    }
+}
