--- SparcFeatures/Pages/_Host.cshtml
@@ -1,30 +1,0 @@
-﻿@page "/"
-@using Microsoft.AspNetCore.Components.Web
-@namespace Sparc.Coop.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"  />
-    <base href="~/" />
-    <link rel="stylesheet" type="text/css" href="https://api.kodekit.io/greffe-winterweight.css">
-    <link href="css/app.css" rel="stylesheet" />
-    <link rel="stylesheet" href="https://use.typekit.net/dda1jal.css">
-    <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
-</head>
-<body>
-    <app>
-        <component type="typeof(Sparc.Coop.App)" render-mode="ServerPrerendered" />
-    </app>
-
-    <script src="_framework/blazor.server.js"></script>
-    <script>
-        function scrollFunction() {
-            const element = document.getElementById("mission");
-            element.scrollIntoView({ behavior: 'smooth' });
-        }
-    </script>
-</body>
-</html>
