+++ SparcFeatures/Routes.razor
@@ -1,0 +1,5 @@
+﻿<h3>Routes</h3>
+
+@code {
+
+}
